Make caddy run as nobody.
Combined with a convention of listening on unprivileged ports,
this means caddy isn't running as root and we don't need to set
special capablities attributes on the executable file.

This works well when caddy isn't expecting to do writes. To
do pre-compiled markdown, or git pulls, other middleware that
does filesystem writes, adjustments to the container fs required.