Doc: README: `s/user/j/g` in /home/user/... paths in configuration examples.
`user` is a keyword for the hub.docker.com markdown variant, so the paths in the examples beginning `/home/user/` had a confusing appearance when shown there (main viewing location) in bold.

TODO(?): change all the host-side paths for the volumes to a sort of best-practice, rather than a /example/path/basically/ as they are now.